Huge refactor for 0.5.0: make namespace and directory structure truly PSR-4; Create EmbedCode entity to not just return a string from this method; Include and implement VideoManagerMeta interface definitions for Video and EmbedCode
