+++ lib/ApiClient/Guzzle5ApiClient.php
