+++ lib/ApiClient/Guzzle6ApiClient.php
