+++ lib/Entity/Channel.php
