+++ lib/Exception.php
