+++ lib/Extractor/TokenExtractor.php
