+++ lib/Factory/Guzzle5ApiClientFactory.php
