+++ lib/Factory/Guzzle6ApiClientFactory.php
