+++ tests/MovingImage/Test/Util/Logging/LoggerAwareTraitImpl.php
@@ -2,7 +2,8 @@
 
 namespace MovingImage\Test\Util\Logging;
 
-use MovingImage\Util\Logging\Traits\LoggerAwareTrait;
+use MovingImage\Client\VMPro\Util\Logging\Traits\LoggerAwareTrait;
+
 
 /**
  * Class LoggerAwareTraitImpl.
